on_wavelink_track_end removed for new inactive_player_timeout parameter. on_wavelink_inactive_player added for inactivity event. !remove and !playnext functions added and tested. old music_cog deleted. swanny_bot.py refactored to reflect new music_cog rename.
