+++ swanny_bot.py
@@ -1,24 +1,23 @@
 import asyncio
-
 import discord
 from discord.ext import commands
-
 import swannybottokens
 from help_cog import help_cog
-#from music_cog import music_cog
 from streamer_cog import streamer_cog
 from special_cog import special_cog
 from video_cog import video_cog
-#from rep_cog import rep_cog
+# from rep_cog import rep_cog
 from twitterfixer_cog import twitterfixer_cog
-from music_cog3 import MusicCog
+from music_cog import MusicCog
 import database
 import wavelink
+
+
 class Swannybot(commands.Bot):
     def __init__(self):
         intents = discord.Intents.default()
         intents.message_content = True
-        intents.presences=True
+        intents.presences = True
         intents.members = True
         super().__init__(
         command_prefix=commands.when_mentioned_or('!'),
@@ -33,11 +32,12 @@ class Swannybot(commands.Bot):
         print(f'Logged in as {self.user} (ID: {self.user.id})')
         print('Bot is ready!')
     async def setup_hook(self) -> None:
-        nodes = [wavelink.Node(uri="http://docker:2333", password=swannybottokens.WavelinkPassword)]
+        nodes = [wavelink.Node(uri="http://docker:2333", password=swannybottokens.WavelinkPassword,
+                               inactive_player_timeout=600)]
         # cache_capacity is EXPERIMENTAL. Turn it off by passing None
         await wavelink.Pool.connect(nodes=nodes, client=self, cache_capacity=100)
         database.dbhandler()
-        await self.load_extension("music_cog3")
+        await self.load_extension("music_cog")
         await self.load_extension("twitterfixer_cog")
         await self.load_extension("video_cog")
         await self.load_extension("help_cog")
